--- a/onmt/decoders/transformer.py
+++ b/onmt/decoders/transformer.py
@@ -115,6 +115,7 @@ class TransformerDecoderLayerBase(nn.Module):
             raise ValueError(f"{layer_norm} layer norm type is not supported")
 
         self.dropout = nn.Dropout(dropout)
+        self.dropout_p = dropout
         self.full_context_alignment = full_context_alignment
         self.alignment_heads = alignment_heads
 
@@ -323,7 +324,8 @@ class TransformerDecoderLayer(TransformerDecoderLayerBase):
         norm_layer_in = self.layer_norm_1(layer_in)
 
         self_attn, _ = self._forward_self_attn(norm_layer_in, dec_mask, step)
-
+        if self.dropout_p > 0:
+            self_attn = self.dropout(self_attn)
         if self.parallel_residual:
             ctx_attn, attns = self.context_attn(
                 enc_out,
@@ -337,16 +339,18 @@ class TransformerDecoderLayer(TransformerDecoderLayerBase):
                 self.feed_forward(norm_layer_in)
                 - norm_layer_in
                 + layer_in
-                + self.dropout(self_attn)
+                + self_attn
                 + ctx_attn
             )
         else:
-            query = self.dropout(self_attn) + layer_in
+            query = self_attn + layer_in
             norm_query = self.layer_norm_2(query)
             ctx_attn, attns = self.context_attn(
                 enc_out, enc_out, norm_query, mask=src_pad_mask, return_attn=return_attn
             )
-            layer_out = self.feed_forward(self.dropout(ctx_attn) + query)
+            if self.dropout_p > 0:
+                ctx_attn = self.dropout(ctx_attn)
+            layer_out = self.feed_forward(ctx_attn + query)
 
         return layer_out, attns
 
@@ -676,7 +680,8 @@ class TransformerLMDecoderLayer(TransformerDecoderLayerBase):
         attn_output, attns = self._forward_self_attn(
             norm_layer_in, dec_mask, step, return_attn=return_attn
         )
-
+        if self.dropout_p > 0:
+            attn_output = self.dropout(attn_output)
         if self.parallel_residual:
             # feed_forward applies residual, so we remove and apply residual with un-normed
             if not self.shared_layer_norm:
@@ -684,11 +689,9 @@ class TransformerLMDecoderLayer(TransformerDecoderLayerBase):
                 ff_in = norm_res_layer_in
             else:
                 ff_in = norm_layer_in
-            layer_out = (
-                self.feed_forward(ff_in) - ff_in + layer_in + self.dropout(attn_output)
-            )
+            layer_out = self.feed_forward(ff_in) - ff_in + layer_in + attn_output
         else:
-            layer_out = self.dropout(attn_output) + layer_in
+            layer_out = attn_output + layer_in
             layer_out = self.feed_forward(layer_out)
 
         return layer_out, attns