--- a/onmt/model_builder.py
+++ b/onmt/model_builder.py
@@ -160,6 +160,9 @@ def load_test_model(opt, device_id=0, model_path=None):
 
     model.eval()
     model.generator.eval()
+    for name, module in model.named_modules():
+        if hasattr(module, "dropout_p"):
+            module.dropout_p = 0.0
     return vocabs, model, model_opt