--- a/onmt/model_builder.py
+++ b/onmt/model_builder.py
@@ -159,7 +159,6 @@ def load_test_model(opt, device_id=0, model_path=None):
     del checkpoint
 
     model.eval()
-    model.generator.eval()
     for name, module in model.named_modules():
         if hasattr(module, "dropout_p"):
             module.dropout_p = 0.0