… onmt/inputters/text_utils.py
… onmt/modules/position_ffn.py
… onmt/modules/util_class.py
… onmt/train_single.py
… onmt/translate/beam_search.py
… onmt/translate/decode_strategy.py
… onmt/translate/translation_server.py
… onmt/translate/translator.py
… onmt/utils/distributed.py
… onmt/utils/misc.py
… onmt/utils/scoring_utils.py
… tools/LM_scoring.py