Many various small optimizations ending up in a huge speed increase:
So while inferencing:
- set dropout_p to zero and avoid going through self.dropout modules
- do not sort / re-sort each minibatch - instead we should sort the bucket and re-sort at the end but for this we need to pass the info of the bucket to translator (future work for more optimization)
- better code when re-building translations

Remove IterOnDevice, instead create tensors directly on the device in tensorify()

Altogether inference speed is ~ x1.5
For the 3003 sentences newstest2014, using a vanilla big transformer on a RTX 4090
end-to-end time with batch size 1001 sentences (3 batches)(fp16): 8.9 seconds
end-to-end time with sames settings ct2 (int8_floatt16): 3.4 seconds

with batch_size 32:
ONMT-py (fp16): 26.5 seconds
CT2 (int8_float16): 17.7 seconds

If we sort upfront the 3003 sentences by length and use batch_size 32, we end up to 18.5 seconds with ONMT-py (hence my point on bucket sorting above)